Update run-data.pl
This script loads the XML files stored in folder 'src', converts the XML trees to triples, and stores all of these in a single RDF graph.The conversion is a very literal one: every node in the XML becomes an entity carrying the same name as in the XML file. The only (initial) step towards 'linkedpolitics' data is in the namespace, which is already set to http://purl.org/linkedpolitics/vocabulary. Further rewriting is done using rewrite_data.pl. When that script is fully tested and ready, the 'rewrite' statement in convert_file can be uncommented to do this all in one go.