Update run-data.pl
Added convert_dir to convert the entire src folder at once.
For each file, the triples are first written to a graph 'buffer' to facilitate the creation of the lpv:text and lpv:speechNo properties. The buffer graph is then copied to the complete graph called 'people' (todo: change name).